Mechanic #4: Really, the intensity of the image does not increase linearly, it increases quickly and then dies down again. I divide the grid into 3 sections and manage the intensity different for each third.
The first third increases in intensity linearly and then last third
decreases linearly.

The middle third is the complicated one that still doesn't quite look
right. I change the deviation property on the random corners a bit and
increase the number of squares at an exponential rate.

My theory is that it would look more accurate if I had more randomness
for the corners on the x axis and a little less on the y